moved masking into its own module for maintainability purposes
